Keeping trust values invariant for baselines
