+++ worlds1/WorldBuilder.py
@@ -14,7 +14,8 @@ from matrx.actions.object_actions import RemoveObject
 from matrx.objects import EnvObject
 from matrx.world_builder import RandomProperty
 from matrx.goals import WorldGoal
-from agents1.OfficialAgent import BaselineAgent, TrustMechanism
+from agents1.OfficialAgent import BaselineAgent
+from beliefs.TrustBelief import TrustMechanism
 from agents1.TutorialAgent import TutorialAgent
 from actions1.CustomActions import RemoveObjectTogether
 from brains1.HumanBrain import HumanBrain
